refactor: improve time-period-selector component (old monthly-button-group)
