+++ app/page.tsx
@@ -6,7 +6,7 @@ import AddTransactionDialog from "./components/add-transaction-dialog";
 import EmptyTransactionHistory from "./components/empty-transaction-history";
 import Card from "./components/financial-summary-card";
 import FinancialTable from "./components/financial-table";
-import MonthlyButtonGroup from "./components/monthly-button-group";
+import TimePeriodSelector from "./components/time-period-selector";
 import { Button } from "./components/ui/button";
 import { ScrollArea } from "./components/ui/scroll-area";
 import Column from "./components/utils/column";
@@ -17,14 +17,17 @@ import { TransactionType } from "./types/transaction-type";
 
 export default function Home() {
   const getCurrentMonth = new Date().getMonth() + 1;
+  const getCurrentYear = new Date().getFullYear();
   const [month, setMonth] = useState<number>(getCurrentMonth);
+  const [year, setYear] = useState<number>(getCurrentYear);
 
   const filteredTransactions = useMemo(() => {
     return mockDataTable.filter((transaction) => {
       const transactionMonth = transaction.date.getMonth() + 1;
-      return transactionMonth === month;
+      const transactionYear = transaction.date.getFullYear();
+      return transactionMonth === month && transactionYear === year;
     });
-  }, [month]);
+  }, [month, year]);
 
   const totalIncoming = useMemo(() => {
     return filteredTransactions.reduce((sum, transaction) => {
@@ -49,7 +52,12 @@ export default function Home() {
   return (
     <Column className="min-h-screen w-full flex bg-[#f4f2ee] items-center sm:justify-center">
       <Column className="items-center space-y-2 w-full max-w-3xl">
-        <MonthlyButtonGroup selectedMonth={month} setSelectedMonth={setMonth} />
+        <TimePeriodSelector
+          selectedYear={year}
+          setSelectedYear={setYear}
+          selectedMonth={month}
+          setSelectedMonth={setMonth}
+        />
         <Flex className="sm:space-x-2 max-sm:space-y-2 flex-col sm:flex-row w-full justify-center items-center">
           <Card title="Entradas" value={totalIncoming} />
           <Card title="Saídas" value={totalOutcoming} />
